Migrate `buffers` into vector-core
This commit migrates `buffers` into vector-core. Resolves #7399. This will allow
us to experiment more quickly with resolutions to #7246. I would like to see
`buffers` as a sub-package to `vector-core` which will mean moving `Event` et al
around, so I intend to do that in a follow-up commit as it'd be quite a dramatic
change.
