+++ lib/vector-core/Cargo.toml
@@ -11,11 +11,14 @@ atomig = { version = "0.2.0", features = ["derive", "serde"] }
 bytes = { version = "1.0.1", default-features = false, features = ["serde"] }
 chrono = { version = "0.4.19", default-features = false, features = ["serde"] }
 dashmap = { version = "4.0.2", default-features = false }
+db-key = { version = "0.0.5", default-features = false }
 derivative = { version = "2.2.0", default-features = false }
 derive_is_enum_variant = { path = "../derive_is_enum_variant" }
+futures = { version = "0.3.14", default-features = false, features = ["compat", "io-compat"], package = "futures" }
 getset = { version = "0.1.1", default-features = false }
 indexmap = { version = "1.6.2", default-features = false, features = ["serde"] }
 lazy_static = { version = "1.4.0", default-features = false }
+leveldb = { version = "0.8.6", default-features = false, optional = true }
 lookup = { path = "../lookup", features = ["arbitrary"] }
 metrics = { version = "0.15.1", default-features = false, features = ["std"]}
 metrics-tracing-context = { version = "0.4.0", default-features = false }
@@ -23,6 +26,7 @@ metrics-util = { version = "0.7.0", default-features = false }
 once_cell = { version = "1.3", default-features = false }
 pest = { version = "2.1.3", default-features = false }
 pest_derive = { version = "2.1.0", default-features = false }
+pin-project = { version = "1.0.7", default-features = false }
 prost = { version = "0.7.0", default-features = false }
 prost-types = { version = "0.7.0", default-features = false }
 regex = { version = "1.5.4", default-features = false, features = ["std", "perf"] }
@@ -31,8 +35,8 @@ serde = { version = "1.0.125", default-features = false, features = ["derive"] }
 serde_json = { version = "1.0.64", default-features = false, features = ["alloc", "raw_value"] }
 serde_yaml = { version = "0.8.17", default-features = false, optional = true }
 shared = { path = "../shared" }
-tokio = { version = "1.5.0", default-features = false, features = ["full"] }
 snafu = { version = "0.6.10", default-features = false, features = ["futures"] }
+tokio = { version = "1.5.0", default-features = false, features = ["full"] }
 toml = { version = "0.5.8", default-features = false }
 tracing = { version = "0.1.26", default-features = false }
 tracing-core = { version = "0.1.17", default-features = false }
@@ -50,12 +54,14 @@ prost-build = "0.7.0"
 criterion = { version = "0.3.4", features = ["html_reports"] }
 quickcheck = "1.0.3"
 pretty_assertions = "0.7.2"
+tokio-test = "0.4.1"
 
 [features]
 default = []
 api = ["async-graphql"]
 lua = ["rlua"]
 vrl = ["vrl-core"]
+disk-buffer = ["leveldb"]
 
 [[bench]]
 name = "lookup"
