+++ lib/vector-core/src/lib.rs
@@ -22,6 +22,7 @@
 
 #[cfg(feature = "api")]
 pub mod api;
+pub mod buffers;
 pub mod config;
 pub mod event;
 pub mod mapping;
